% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pay_plot.R
\name{avg_pay_plt}
\alias{avg_pay_plt}
\title{Box plot of Average Medicare Payment vs DRUG code}
\usage{
avg_pay_plt(xcode)
}
\arguments{
\item{xcode}{input a 3 digit drug code}
}
\value{
a box plot of Average Medicare Payment vs DRUG code
}
\description{
This function produces a box plot of average medicare payment with a 3 digit drug code
requires ipps data to be loaded
}
\examples{
avg_pay_plot(039)
}
