% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pay_stat.R
\name{stat_avg_pay}
\alias{stat_avg_pay}
\title{Measure statistic function
This function calculates a measure statistic with average medicare payments
requires ipps_data to be loaded}
\usage{
stat_avg_pay(x)
}
\arguments{
\item{x}{a measure statistic such as mean, meadian, sd etc.}
}
\value{
a measure statistic
}
\description{
Measure statistic function
This function calculates a measure statistic with average medicare payments
requires ipps_data to be loaded
}
\examples{
stat_avg_pay(mean)
}
